requirements.txt
mysql-connector-python==8.3.0
numpy==1.24.3
pandas==2.2.0
setuptools==69.0.3
streamlit==1.31.0
python-dotenv==1.0.1

